requirements.txt
pyserial
pyvisa
pydmm
paho-mqtt